make life cycle hook private (for now)
